Fix for tabUI do-not-disturb setting against notification-popup
In honeycomb this will kill the nagview, silence the notifications-sounds
and transform notification incons into bars (- - - -), which is really silly
actually.
this fix will make it kill the nagscreen and thats that. good for all google
approved devices because it wont overlay our buttons anymore.

Patchset 2: Fix whitespaces
Patchset 3: Indenting

Change-Id: I3a1803a0aad53cba16d33205e6fca3756d1470ab