sysUI: fix for double recents and foreground activity
1. aosp has a rare bug that displays twice the amount of
recents at startup - sometimes. its hard to reproduce because it is
caused by the interaction of two concurrent threads. one of these
will try to cancel the other, but in some cases that flag comes in
too late and the result is double-recents.

2. google blinds out the foreground task. this will unhide it.
its actually great for quitting an app or for accessing hybrid props.

Change-Id: Ia9fd6479d89ea0dfe14e4d83697a01f1b3931aa3