sysUI: double ime toggle + compat button
tabUI's ime toggle is better no doubt, but it kind of interferes with
the way standard aosp handles it, and it is hardcoded. for better or for
worse, this will kill it. the compatbutton aswell.

the result is, we get one ime toggle in the form of a standard notification
icon. this icon has no function, unless we swipe up the center and tap it,
just like in phoneUI.

Change-Id: I4114d4f9b22a7ceac70a8bb4ce9ccffb08ea2ded