Restore Quick Pulldown & Correct Brightness tile.
Commit removes the inclusion of the SwitchToSettings function from the NotificationPanelView to restore the correct Quick Pulldown functionality. SwitchToSettings function remains in the code until I can fully implement this whilst allow full and correct function of the Quick Pulldown.

Modification of the BrightnessTile to correct non-functioning behaviour, prior to this commit the Brightness Tile made no change to the screen brightness. The tile now functions correctly

Change-Id: I83988411defef9e277dcdf95c9a360358b6fd357