QS: Enable ProfileTile (1/2)
Enables the ProfileTile to control the Profile currently in use.

This was partially implemented but had various parts missing, due to QuickSettings being overhauled I was unable to cherry-pick the original CM addition of the ProfileTile as it was originally authored under the old QuickSettings implementation.

Compiled and tested with no issues, changes profile correctly.

Change-Id: Ib06524a29ef49f48af88b937595b926af6f3f99f