profiles: Broadcast profile updates
Add broadcast to indicate profile changes that does not
cause a profile to be selected.

Use the new broadcast to update the Profile Tile more
reliably.

Currently the tile update is triggered by a profile
selection. However, the active profile may change without
triggering a selection. This happens, for instance, on
locale changes.

The Profile Tile does handle locale changes, but is subject
to a race condition. If it handles the locale change before
the profile manager does, the name change will not be seen.

Conflicts:

	packages/SystemUI/src/com/android/systemui/quicksettings/ProfileTile.java

Change-Id: Ifac462bde348717d4d1526e0c6688212024eeef3