Forward port of lock before unlock setting [1/2]
This is a work in progress, do not merge yet.
This feature will allow users that use secure lockscreen to choose wether
they want to see the generic lockscreen before the secure lockscreen allowing
them to make use of other lockscreen features.

ps2: Moved to Secured Settings,
deleted an extra define of lock_before_unlock
deleted extra strings that were tagging along with the commits.

ps3: was a fail :(

Change-Id: Ifb33fffbe1930128a7436b21247a7e17d3dd8e1d