WMS: fix no thumbnail image of Gallery in the recent tasks list.
When it takes the screenshot of an application, it will calculate
the minLayer and maxLayer first, and if a window is full screen,
it won't deal with the windows below it. However, the surfaceview
is placed below a window but will be displayed at front, this make
the surface view can not be included when take the screenshot of
an application, then it diaplays a black area in the recent tasks
list.

If the currect window is full screen, no need to finish the
calculation.

CR-Fixed:566752

Change-Id: Iaf3e9a3e6520007571ea545ccc9c05d4b1b416fd