Add checkbox in recents settings to make app shortcuts in recents optional
Since some time there is a shortcut bar in recents for some selected apps.
I don't need it and it clutters the interface. Maybe somebody else likes
the shortcuts so I thought I'd just make it optional.
This is the UI part. Other part is in packages/apps/Settings with
Change-Id: I0f365616e048292e43552bcf84bf24798d6432b9
This patch needs the before mentioned patch.
I didn't find out how to make the patches depend on each other.