Unifying the Pac Image.
Some unicorns got in a made a mess of the stables, I sent someone over
to sort it out with no luck. But I've got this stuff here to fix up most
the issues he created.

AND why do we have stables if we don't have any horses! It's just asking
for unicorn problems!

Change-Id: Ifbaa7a4ebde18adb6fcf8213778be475d67f5796