fb: Handle navbar heights correct on dpi changes + more (1/2)
Till now we set the height in px in the settings. Which is a problem
especially on multiDPI roms like we are when the user afterwards changes
the dpi due that it isn't calculated correct anymore on the new dpi.

This commit does
- save the dp value in provider and calculate the real px in fb
- clean up the settings stuff for it + do it a bit nicer (only settings part
  where we do not set the value for the listview)
- as well this commit disables movable navbar for tablets (we may will enable
  it in future but it needs a lot more work to have it on tablets correct working)
- add Converter class for common methods we always use....till now only px->dp
  and dp->px method
----------------------------------------------------------------------------------
Conflicts:
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java

PS2: Add DeviceUtils
PS3: Add Landscape height
PS4: Remove Landscape height

Change-Id: I3ddcb993f82758bf2d7a183dcb8023b2fcb8f6c8