Prevent a duplicated registration of Observers
If same Observer and same URI are registered, the callback functions are called multiple times
when data identified by a given content URI changes.
It makes system slow and it cause ANR sometimes.
And When run unregisterContentObserver(), just one deleted
but others are remained in mObservers.
So add prevent code to avoid to duplicated registraion of Observers

Change-Id: If92e861080d9d361db62f32adf2f4c3d1bb14f27
Signed-off-by: muzbit.kim <muzbit.kim@lge.com>