Prevent Showing the Crash Dialog for System UI
After system ui crashes, the crash dialog can prevent the keyguard from
showing if the screen was shut off before the dialog was dismissed. The
user can't recover from this without waiting for the dialog to time out
which takes several minutes.

Not showing the dialog for system ui is a safe way of preventing this
from occuring.

Change-Id: Iac235f3af244f735ef0e043098ed84874e3a84f9