Framework: Make the dialog divider color themeable
The holo blue color might not be suitable for some themes. This commit
swaps out the background of all titleDivider elements to use another
resource. This has no difference from the perspective of the user of a
"stock" build but this is huge for theme users.

Thanks to Arz Bhatia for the suggestion.
----------------------------------------
PS1: PAC Upload, commented out those changes.

Conflicts:
	core/res/res/layout-xlarge/activity_list.xml
	core/res/res/layout/alert_dialog_holo_dark.xml
	core/res/res/layout/dialog_custom_title_holo.xml
	core/res/res/layout/dialog_title_holo.xml
	core/res/res/layout/dialog_title_icons_holo.xml

Change-Id: Icb917220accac3ad38770067120a7b1be894961a