SystemUI: Fix SearchPanelView miscalculation
The SystemUI declares the navbar to be movable to the side in landscape when the short
side's DP calculation equals to less than 600. This means that the nav bar will be at the
bottom/the screen is considered "large" if it's greater OR EQUAL to 600, not just greater.

Fixes OMNI-1129 and OMNI-1130.

Change-Id: I90322e40ebf0c4bb94146e28eb2d9b8f8a4444f8