QuietHours: don't schedule alarms in the past
On the last day of the month, the trigger-alarm for the next day
was actually scheduled on the first day of the current month, which
caused it to fire immediately and then repeating this, leading
to high battery drain.

Fix this by scheduling the next alarm on the real next day.

Change-Id: Ied3b1dd3101650842a798d7e00e2a3a794ac8639