Select correct 12 or 24 time format according to Locale
apps that depend on java.text.DateFormat and friends to format
a date or time string will expect it to correctly handle 12 or 24
time format according to the current locale and user selections in
settings. This fixes a case where this was not being done properly
when setting up a process in a locale that defaults to 24 hour
time display when the user had not explicitly overriden the
default format. It also includes a fix to update time formats
if the user changes their locale without requiring the process
to restart.

This is a fix for CYNGNOS-495 which noted incorrect date
formatting in the call log after choosing a locale that uses
24 hour time formats.

Change-Id: I449e7ef22f09a29de775aaf3fd8d38ca634a0bb4