SystemUI: Allow clearing recents tasks within recents button
* Forward port and simplify code logic for lollipop.
* Update recents button properly every time setNavigationIconHints() is
called. Remove the callback that is no longer really needed.
* Tasks will be cleared within long-pressing recents. It will not clear
tasks when accessibility mode and/or screen pinning is in use, long-
press
event will be treathed then as a normal recent to allow unlocking
normally screen-pinning.
* Defaulted to clearing all of them, on the spot will ask to keep last
task.
* Drawables by arzbhatia
* Thanks to TheMrcool212 for the initial port patch and for adapting
cm methods used to dismiss tasks, ignoring last app as for kk.

* Reworked drawables
* Fixed recents button not updating correctly
* Ported to CyanogenMod 13

Signed-off-by: BigBrother1984 <carlosavignano@aospa.co>
Change-Id: I0a793fc8a174d956ff9b4b9c5d3259897c439795
Signed-off-by: lion0738 <lion0738@naver.com>