SysUI: Check swiping progress before collapsing KBA
Collapsing the KeyguardBottomArea while a swipe is in progress causes
the window to collapse to a smaller size and cut off the circular
animations from the affordance views.  Luckliy there is a method we
can call to see if we are in such a state, so let's use it.

Change-Id: I45241ed17c3c90580fb04845700f5632b34db187
TICKET: CYNGNOS-2824